moved WorkflowProcess to de.uka.ipd.sdq.workflow
