+++ bundles/de.uka.ipd.sdq.workflow.ui/META-INF/MANIFEST.MF
@@ -4,10 +4,7 @@ Bundle-Name: Workflow UI
 Bundle-SymbolicName: de.uka.ipd.sdq.workflow.ui;singleton:=true
 Bundle-Version: 5.2.1.qualifier
 Export-Package: de.uka.ipd.sdq.workflow.ui
-Require-Bundle: de.uka.ipd.sdq.workflow.logging.console;bundle-version="1.0.0",
- org.eclipse.core.runtime;bundle-version="3.5.0",
- org.eclipse.debug.core;bundle-version="3.5.0",
- de.uka.ipd.sdq.workflow;bundle-version="1.1.0",
+Require-Bundle: de.uka.ipd.sdq.workflow;bundle-version="1.1.0",
  de.uka.ipd.sdq.dialogs;bundle-version="2.0.1",
  org.eclipse.ui;bundle-version="3.5.0"
 Bundle-Vendor: Palladio Project
